requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
Django==2.2.1
djangorestframework==3.9.4
feedparser==5.2.1
kombu==4.5.0
pytz==2019.1
sqlparse==0.3.0
vine==1.3.0
